Move contents of .h files into .cc files for plugins
