+++ RecoBTag/SoftLepton/plugins/SoftLepton.cc
@@ -8,6 +8,7 @@
  Description: CMSSW EDProducer for soft lepton b tagging.
 
  Implementation:
+     The actual tagging is performed by SoftLeptonAlgorithm.
 */
 
 // Original Author:  fwyzard
@@ -17,9 +18,36 @@
 #include <string>
 #include <utility>
 #include <cmath>
+#include <map>
+
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/stream/EDProducer.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
+#include "DataFormats/Common/interface/RefToBase.h"
+#include "DataFormats/Math/interface/Vector3D.h"
+#include "DataFormats/GeometryVector/interface/GlobalVector.h"
+#include "DataFormats/JetReco/interface/Jet.h"
+#include "DataFormats/JetReco/interface/JetTracksAssociation.h"
+#include "DataFormats/VertexReco/interface/VertexFwd.h"
+#include "DataFormats/VertexReco/interface/Vertex.h"
+#include "DataFormats/Common/interface/ValueMap.h"
+#include "DataFormats/EgammaCandidates/interface/GsfElectron.h"
+#include "DataFormats/EgammaCandidates/interface/Electron.h"
+#include "DataFormats/ParticleFlowCandidate/interface/PFCandidateFwd.h"
+#include "DataFormats/ParticleFlowCandidate/interface/PFCandidate.h"
+#include "DataFormats/TrackReco/interface/Track.h"
+#include "DataFormats/MuonReco/interface/Muon.h"
+#include "DataFormats/MuonReco/interface/MuonSelectors.h"
+#include "DataFormats/BTauReco/interface/SoftLeptonTagInfo.h"
+#include "TrackingTools/Records/interface/TransientTrackRecord.h"
 
 #include "FWCore/Utilities/interface/InputTag.h"
 #include "FWCore/ParameterSet/interface/ParameterSetDescription.h"
+
+#include "FWCore/Framework/interface/ModuleFactory.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
 
 // ROOT::Math vectors (aka math::XYZVector)
 #include "DataFormats/Math/interface/LorentzVector.h"
@@ -33,7 +61,73 @@
 #include "TrackingTools/TransientTrack/interface/TransientTrackBuilder.h"
 #include "TrackingTools/Records/interface/TransientTrackRecord.h"
 #include "TrackingTools/IPTools/interface/IPTools.h"
-#include "SoftLepton.h"
+
+class SoftLepton : public edm::stream::EDProducer<> {
+public:
+  explicit SoftLepton(const edm::ParameterSet &iConfig);
+  ~SoftLepton() override;
+  static void fillDescriptions(edm::ConfigurationDescriptions &descriptions);
+
+  struct TrackCompare {
+    inline bool operator()(const edm::RefToBase<reco::Track> &t1, const edm::RefToBase<reco::Track> &t2) const {
+      return t1.key() < t2.key();
+    }
+  };
+
+  typedef std::map<unsigned int, float> LeptonIds;
+  typedef std::map<edm::RefToBase<reco::Track>, LeptonIds, TrackCompare> Leptons;
+
+  // generic interface, using a TrackRefVector for lepton tracks
+  reco::SoftLeptonTagInfo tag(const edm::RefToBase<reco::Jet> &jet,
+                              const reco::TrackRefVector &tracks,
+                              const Leptons &leptons,
+                              const reco::Vertex &primaryVertex) const;
+
+protected:
+  // generic interface, using a TrackRefVector for lepton tracks
+
+  GlobalVector refineJetAxis(const edm::RefToBase<reco::Jet> &jet,
+                             const reco::TrackRefVector &tracks,
+                             const edm::RefToBase<reco::Track> &exclude = edm::RefToBase<reco::Track>()) const;
+
+  static double relativeEta(const math::XYZVector &vector, const math::XYZVector &axis);
+
+  static double boostedPPar(const math::XYZVector &vector, const math::XYZVector &axis);
+
+private:
+  void produce(edm::Event &event, const edm::EventSetup &setup) override;
+
+  // configuration
+  const edm::InputTag m_jets;
+  const edm::EDGetTokenT<reco::JetTracksAssociationCollection> token_jtas;
+  const edm::EDGetTokenT<edm::View<reco::Jet> > token_jets;
+  const edm::EDGetTokenT<reco::VertexCollection> token_primaryVertex;
+  const edm::InputTag m_leptons;
+  const edm::EDGetTokenT<edm::View<reco::GsfElectron> > token_gsfElectrons;
+  const edm::EDGetTokenT<edm::View<reco::Electron> > token_electrons;
+  const edm::EDGetTokenT<reco::PFCandidateCollection> token_pfElectrons;
+  const edm::EDGetTokenT<edm::View<reco::Muon> > token_muons;
+  const edm::EDGetTokenT<edm::View<reco::Track> > token_tracks;
+  const edm::InputTag m_leptonCands;
+  const edm::EDGetTokenT<edm::ValueMap<float> > token_leptonCands;
+  const edm::InputTag m_leptonId;
+  const edm::EDGetTokenT<edm::ValueMap<float> > token_leptonId;
+
+  // service used to make transient tracks from tracks
+  const edm::ESGetToken<TransientTrackBuilder, TransientTrackRecord> token_builder;
+  const TransientTrackBuilder *m_transientTrackBuilder;
+
+  // algorithm configuration
+  const unsigned int m_refineJetAxis;
+  const double m_deltaRCut;
+  const double m_chi2Cut;
+
+  // specific for reco::Muons
+  const muon::SelectionType m_muonSelection;
+
+  // nominal beam spot position
+  static const reco::Vertex s_nominalBeamSpot;
+};
 
 enum AxisType {
   AXIS_CALORIMETRIC = 0,              // use the calorimietric jet axis
@@ -441,3 +535,5 @@ void SoftLepton::fillDescriptions(edm::ConfigurationDescriptions &descriptions) 
   desc.add<double>("leptonChi2Cut", 9999.0);
   descriptions.addDefault(desc);
 }
+
+DEFINE_FWK_MODULE(SoftLepton);
