+++ RecoBTag/SoftLepton/plugins/SoftPFElectronTagInfoProducer.cc
@@ -1,4 +1,5 @@
-#include "RecoBTag/SoftLepton/plugins/SoftPFElectronTagInfoProducer.h"
+#include "FWCore/Framework/interface/ModuleFactory.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
 #include "DataFormats/EgammaCandidates/interface/GsfElectron.h"
 #include "DataFormats/EgammaCandidates/interface/GsfElectronFwd.h"
 #include "DataFormats/GsfTrackReco/interface/GsfTrack.h"
@@ -15,15 +16,64 @@
 #include "DataFormats/ParticleFlowCandidate/interface/PFCandidateFwd.h"
 #include "DataFormats/BTauReco/interface/SoftLeptonTagInfo.h"
 
+#include "FWCore/Framework/interface/stream/EDProducer.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "DataFormats/EgammaCandidates/interface/GsfElectron.h"
+#include "DataFormats/Common/interface/ValueMap.h"
+#include "DataFormats/ParticleFlowCandidate/interface/PFCandidate.h"
+#include "DataFormats/JetReco/interface/PFJetCollection.h"
+
+#include "TrackingTools/TransientTrack/interface/TransientTrack.h"
+#include "TrackingTools/TransientTrack/interface/TransientTrackBuilder.h"
+#include "TrackingTools/Records/interface/TransientTrackRecord.h"
+
+// Vertex
+#include "DataFormats/VertexReco/interface/Vertex.h"
+#include "DataFormats/VertexReco/interface/VertexFwd.h"
+#include "DataFormats/BTauReco/interface/SoftLeptonTagInfo.h"
+#include "DataFormats/PatCandidates/interface/Electron.h"
+
 // Transient Track and IP
 #include "TrackingTools/TransientTrack/interface/TransientTrack.h"
 #include "TrackingTools/TransientTrack/interface/TransientTrackBuilder.h"
 #include "TrackingTools/Records/interface/TransientTrackRecord.h"
 #include "TrackingTools/IPTools/interface/IPTools.h"
 #include "DataFormats/GeometryVector/interface/GlobalVector.h"
-#include <cmath>
 #include "CommonTools/Egamma/interface/ConversionTools.h"
 #include "DataFormats/PatCandidates/interface/Electron.h"
+
+#include <cmath>
+#include <vector>
+
+// SoftPFElectronTagInfoProducer:  the SoftPFElectronTagInfoProducer takes
+// a PFCandidateCollection as input and produces a RefVector
+// to the likely soft electrons in this collection.
+
+class SoftPFElectronTagInfoProducer : public edm::stream::EDProducer<> {
+public:
+  SoftPFElectronTagInfoProducer(const edm::ParameterSet& conf);
+  ~SoftPFElectronTagInfoProducer() override;
+
+private:
+  void produce(edm::Event&, const edm::EventSetup&) override;
+  bool isElecClean(edm::Event&, const reco::GsfElectron*);
+  float boostedPPar(const math::XYZVector&, const math::XYZVector&);
+
+  // service used to make transient tracks from tracks
+  const TransientTrackBuilder* transientTrackBuilder;
+  edm::EDGetTokenT<reco::VertexCollection> token_primaryVertex;
+  edm::EDGetTokenT<edm::View<reco::Jet> > token_jets;
+  edm::EDGetTokenT<edm::View<reco::GsfElectron> > token_elec;
+  edm::EDGetTokenT<reco::BeamSpot> token_BeamSpot;
+  edm::EDGetTokenT<reco::ConversionCollection> token_allConversions;
+  const edm::ESGetToken<TransientTrackBuilder, TransientTrackRecord> token_builder;
+  float DeltaRElectronJet, MaxSip3Dsig;
+  bool goodvertex;
+
+  const reco::Vertex* vertex;
+};
 
 SoftPFElectronTagInfoProducer::SoftPFElectronTagInfoProducer(const edm::ParameterSet& conf)
     : token_builder(esConsumes(edm::ESInputTag("", "TransientTrackBuilder"))) {
@@ -148,3 +198,5 @@ float SoftPFElectronTagInfoProducer::boostedPPar(const math::XYZVector& vector, 
   ROOT::Math::BoostX boost(-jet.Beta());
   return boost(lepton).x();
 }
+
+DEFINE_FWK_MODULE(SoftPFElectronTagInfoProducer);
