+++ RecoBTag/SoftLepton/plugins/module.cc
@@ -1,10 +1,6 @@
 #include "FWCore/Framework/interface/ModuleFactory.h"
 #include "FWCore/Framework/interface/MakerMacros.h"
 
-#include "RecoBTag/SoftLepton/plugins/SoftLepton.h"
-#include "RecoBTag/SoftLepton/plugins/SoftPFElectronTagInfoProducer.h"
-#include "RecoBTag/SoftLepton/plugins/SoftPFMuonTagInfoProducer.h"
-
 #include "DataFormats/EgammaCandidates/interface/GsfElectron.h"
 #include "RecoBTag/SoftLepton/interface/GenericSelectorByValueMap.h"
 
@@ -15,10 +11,6 @@
 #include "RecoBTag/SoftLepton/interface/ElectronTagger.h"
 #include "RecoBTag/SoftLepton/interface/MuonTagger.h"
 #include "RecoBTag/SoftLepton/interface/MuonTaggerNoIP.h"
-
-DEFINE_FWK_MODULE(SoftLepton);
-DEFINE_FWK_MODULE(SoftPFElectronTagInfoProducer);
-DEFINE_FWK_MODULE(SoftPFMuonTagInfoProducer);
 
 // "float" is the type stored in the ValueMap
 typedef edm::GenericSelectorByValueMap<reco::GsfElectron, float> BtagGsfElectronSelector;
